Refactor `collada.py` and `geometry.py`
