+++ src/tqec/__init__.py
@@ -21,6 +21,8 @@ from .computation import (
     ZXKind,
     ZXEdge,
     ZXGraph,
+)
+from .interop import (
     display_collada_model,
     read_block_graph_from_dae_file,
     write_block_graph_to_dae_file,
