+++ src/tqec/computation/__init__.py
@@ -2,8 +2,3 @@ from tqec.computation.zx_graph import ZXKind, ZXEdge, ZXGraph, ZXNode
 from tqec.computation.cube import Cube, CubeKind, ZXCube, Port, YCube
 from tqec.computation.pipe import Pipe, PipeKind
 from tqec.computation.block_graph import BlockGraph, BlockKind
-from tqec.computation.collada import (
-    display_collada_model,
-    read_block_graph_from_dae_file,
-    write_block_graph_to_dae_file,
-)
