+++ src/tqec/interop/__init__.py
@@ -1,0 +1,5 @@
+from tqec.interop.collada import (
+    read_block_graph_from_dae_file,
+    write_block_graph_to_dae_file,
+)
+from tqec.interop.collada_html_viewer import display_collada_model
