+++ src/tqec/position.py
@@ -91,6 +91,15 @@ class Position3D:
         """Shift the position by the given offset."""
         return Position3D(self.x + dx, self.y + dy, self.z + dz)
 
+    def shift_in_direction(self, direction: Direction3D, shift: int) -> Position3D:
+        """Shift the position in the given direction by the given shift."""
+        if direction == Direction3D.X:
+            return self.shift_by(dx=shift)
+        elif direction == Direction3D.Y:
+            return self.shift_by(dy=shift)
+        else:
+            return self.shift_by(dz=shift)
+
     def is_neighbour(self, other: Position3D) -> bool:
         """Check if the other position is near to this position, i.e. Manhattan
         distance is 1."""
@@ -124,3 +133,42 @@ class Direction3D(Enum):
 
     def __str__(self) -> str:
         return self.name
+
+
+@dataclass(frozen=True)
+class SignedDirection3D:
+    """Signed directions in the 3D spacetime diagram."""
+
+    direction: Direction3D
+    towards_positive: bool
+
+    def __neg__(self) -> SignedDirection3D:
+        return SignedDirection3D(self.direction, not self.towards_positive)
+
+
+@dataclass(frozen=True, order=True)
+class FloatPosition3D:
+    """A 3D float position."""
+
+    x: float
+    y: float
+    z: float
+
+    def shift_by(self, dx: float = 0, dy: float = 0, dz: float = 0) -> FloatPosition3D:
+        """Shift the position by the given offset."""
+        return FloatPosition3D(self.x + dx, self.y + dy, self.z + dz)
+
+    def shift_in_direction(
+        self, direction: Direction3D, shift: float
+    ) -> FloatPosition3D:
+        """Shift the position in the given direction by the given shift."""
+        if direction == Direction3D.X:
+            return self.shift_by(dx=shift)
+        elif direction == Direction3D.Y:
+            return self.shift_by(dy=shift)
+        else:
+            return self.shift_by(dz=shift)
+
+    def as_tuple(self) -> tuple[float, float, float]:
+        """Return the position as a tuple."""
+        return astuple(self)
